pyproject.toml
[project]
name = "sam3"
version = "0.0.1"
description = "fun with SAM3 model"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.4",
    "huggingface-hub>=1.1.4",
    "jinja2>=3.1.4",
    "matplotlib>=3.10.7",
    "pillow>=12.0.0",
    "protobuf>=6.33.1",
    "python-multipart>=0.0.12",
    "requests>=2.32.5",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "uvicorn>=0.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sam3"]
